Non breaking changes
* Moved resolvers into its own file
* Changes to userId use
