+++ .env.example
@@ -3,4 +3,4 @@ ENDPOINT=http://localhost:8080/v1/graphql
 HASURA_ADMIN_SECRET=myadminsecretkey
 
 # Auth
-JWT_SECRET=pleasechangeme
+JWT_SECRET=aqjQrhtpDqjBpXGAfGFMG43GZnUZX8bH
